+++ mealkitary-api/src/test/kotlin/com/mealkitary/WebIntegrationTestSupport.kt
@@ -11,12 +11,12 @@ import com.mealkitary.reservation.application.port.input.GetReservationQuery
 import com.mealkitary.reservation.application.port.input.PayReservationUseCase
 import com.mealkitary.reservation.application.port.input.RejectReservationUseCase
 import com.mealkitary.reservation.application.port.input.ReserveProductUseCase
-import com.mealkitary.shop.adapter.input.web.GetProductController
-import com.mealkitary.shop.adapter.input.web.GetReservableTimeController
-import com.mealkitary.shop.adapter.input.web.GetShopController
 import com.mealkitary.shop.application.port.input.GetProductQuery
 import com.mealkitary.shop.application.port.input.GetReservableTimeQuery
 import com.mealkitary.shop.application.port.input.GetShopQuery
+import com.mealkitary.shop.web.GetProductController
+import com.mealkitary.shop.web.GetReservableTimeController
+import com.mealkitary.shop.web.GetShopController
 import com.ninjasquad.springmockk.MockkBean
 import io.kotest.core.spec.style.AnnotationSpec
 import io.kotest.extensions.spring.SpringExtension
