+++ mealkitary-api/src/test/kotlin/com/mealkitary/shop/web/GetProductControllerTest.kt
@@ -1,4 +1,4 @@
-package com.mealkitary.shop.adapter.input.web
+package com.mealkitary.shop.web
 
 import com.mealkitary.WebIntegrationTestSupport
 import com.mealkitary.common.exception.EntityNotFoundException
